pyproject.toml
[project]
name = "python-practice"
version = "1.0"
description = "python-practice"
authors = [
    {name = "birdkyle7918",email = "bbirdkkyle@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.12.15)",
    "aiosignal (==1.4.0)",
    "anyio (==4.10.0)",
    "attrs (==25.3.0)",
    "black (==25.1.0)",
    "blinker (==1.9.0)",
    "build (==1.3.0)",
    "cachecontrol (==0.14.3)",
    "cerberus (==1.3.7)",
    "certifi (==2025.8.3)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.3)",
    "cleo (==2.1.0)",
    "click (==8.2.1)",
    "contourpy (==1.3.3)",
    "crashtest (==0.4.1)",
    "cycler (==0.12.1)",
    "dateparser (==1.2.2)",
    "dephell-archive (==0.1.7)",
    "dephell-argparse (==0.1.3)",
    "dephell-changelogs (==0.0.1)",
    "dephell-discover (==0.2.10)",
    "dephell-licenses (==0.1.7)",
    "dephell-links (==0.1.5)",
    "dephell-markers (==1.0.3)",
    "dephell-pythons (==0.1.15)",
    "dephell-setuptools (==0.2.5)",
    "dephell-shells (==0.1.5)",
    "dephell-specifier (==0.3.0)",
    "dephell-venvs (==0.1.18)",
    "dephell-versioning (==0.1.2)",
    "distlib (==0.4.0)",
    "docutils (==0.22)",
    "dulwich (==0.22.8)",
    "fastjsonschema (==2.21.2)",
    "filelock (==3.19.1)",
    "findpython (==0.6.3)",
    "flask (==3.1.1)",
    "fonttools (==4.59.0)",
    "frozenlist (==1.7.0)",
    "gunicorn (==23.0.0)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "installer (==0.7.0)",
    "itsdangerous (==2.2.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.3.0)",
    "jieba (==0.42.1)",
    "jinja2 (==3.1.6)",
    "keyring (==25.6.0)",
    "kiwisolver (==1.4.9)",
    "m2r (==0.3.1)",
    "markupsafe (==3.0.2)",
    "matplotlib (==3.10.5)",
    "mistune (==0.8.4)",
    "more-itertools (==10.8.0)",
    "msgpack (==1.1.1)",
    "multidict (==6.6.4)",
    "mypy-extensions (==1.1.0)",
    "mysql-connector-python (==9.4.0)",
    "numpy (==2.3.2)",
    "packaging (==25.0)",
    "pathspec (==0.12.1)",
    "pbs-installer (==2025.9.2)",
    "pexpect (==4.9.0)",
    "pillow (==11.3.0)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.8)",
    "poetry (==2.1.4)",
    "poetry-core (==2.1.3)",
    "propcache (==0.3.2)",
    "ptyprocess (==0.7.0)",
    "pyarrow (==21.0.0)",
    "pycparser (==2.22)",
    "pyparsing (==3.2.3)",
    "pyproject-hooks (==1.2.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-telegram-bot (==22.3)",
    "pytz (==2025.2)",
    "rapidfuzz (==3.14.0)",
    "redis (==6.4.0)",
    "regex (==2025.7.34)",
    "requests (==2.32.4)",
    "requests-toolbelt (==1.0.0)",
    "ruamel-yaml (==0.18.15)",
    "ruamel-yaml-clib (==0.2.12)",
    "scipy (==1.16.1)",
    "setuptools (==80.9.0)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "tencentcloud-sdk-python (==3.0.1441)",
    "termcolor (==2.3.0)",
    "tomlkit (==0.13.3)",
    "tqdm (==4.67.1)",
    "trove-classifiers (==2025.8.26.11)",
    "tzlocal (==5.3.1)",
    "urllib3 (==2.5.0)",
    "virtualenv (==20.32.0)",
    "werkzeug (==3.1.3)",
    "wheel (==0.45.1)",
    "xattr (==1.2.0)",
    "yarl (==1.20.1)",
    "yaspin (==3.1.0)",
    "zstandard (==0.24.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pymysql (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
